Отрисуй меня полностью (часть 2). Третья проверка
