+++ index.html
@@ -230,14 +230,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <!-- Шаблон комментария под полноразмерным изображением -->
-  <template id="comment">
-    <li class="social__comment">
-      <img class="social__picture" src="" alt="" width="35" height="35">
-       <p class="social__text"></p>
-    </li>
-  </template>
-
   <script src="js/main.js" type="module"></script>
 </body>
 </html>
