+++ js/main.js
@@ -1,7 +1,7 @@
 import { createImages } from './data.js';
 import { createPreviews } from './create-previews.js';
-import { initBigPicture } from './init-big-picture.js';
+import { closeBigPicture } from './close-big-picture.js';
 
 const images = createImages();
 createPreviews(images);
-initBigPicture(images);
+closeBigPicture();
